Slettet passwordverify mappen (det ligger i functionality)
