+++ CDIO1/src/functionality/IFunctionality.java
@@ -8,6 +8,6 @@ public interface IFunctionality {
     
     boolean isUserIDPresent(int ID, int[] IDs);
     int[] getUserIDs(List<UserDTO> users);
-    boolean verifyPassword(String pass) throws Exception;
+    boolean verify(UserDTO user, String pass) throws Exception;
     
 }
